Unified output for Zabbix or other machine-parsing integrations (#56)
* Add --no-scale argument for bare number output

Added a command line argument `--no-scale` which removes the graphs and
all scale decorators for the temperatere output.  This is useful when
parsing the numbers mechanically for use with system monitoring or other
automated tasks which want to view the current temperatures.

* Unify Printer.format_* functions

Made the Printer global settings uniform and created wrapper format_
function to toggle the display of labels on and off based on the
setting.

Also modified the usage help to be proper Backus-Naur syntax.

* Support display arguments for all outputs

Alter the bare puts commands to support the new display arguments for
fan speed and other items (not just temperature).  Battery is untested
at this point because I'm on a Mac Pro with no battery.

* Unify all line value outputs with print_item_line() function

Simply outputting a key/vale line with a helper proc in the Printer
class so the majority of the `puts` output is all done from the same
place.

* Force numeric value when generating sparkline

* Value and threshholds need to agree with unified print

* Drop sparklines for charge for now

* Add support for reverse-logic sparklines

If the thresholds array passed to gen_sparkline contains negative
values, treat 100% as the best-case instead of the worst-case.

* Clean up batter charge and capacity display

* Wrong variable name for scan

* Add zabbix discovery scan style

* Zabbix needs perfectly-formed JSON for the discovery

That trailing comma is enough to confuse Zabbix.  Re-work the 'scan
zabbix' code so that the data list is well-formed.

* Use descending thresholds instead of negative values

It seems more logical to detect a reverse threshold sparkline if the
values in the array are descending, instead of the goofy negative value
technique I used before.

* Document Zabbix integration in usage and README

Added minimal description of the Zabbix integration to the README and
command usage help text.

* Move abs_thresholds to Utils and fix whitespace

* Use printf formatting instead of tabs for columns

* Revert to original Design capacity label

* Use Utils abs_thresholds for cycle count graph

* Whitespace cleanup

* Use the result we just calculated

* Two spaces are needed here for some reason

* Documentation and code comment improvements

* Replace bare puts with JSON.generate for zabbix output

* Missing braces from Zabbix data structure